refactor(server): extract session reaper and config watcher to services/server (#4243 step 2)
- session-reaper.ts: reapStaleSessions + reapZombieSessions + zombie constants
- config-watcher.ts: setupConfigWatcher + handleConfigReload
- server.ts: 840 → 672 lines (-20%)
- Zero behavior changes, all 5804 tests pass
